A few more renaming and modelling (moves code to a single file)
